Refactor download logic (and make 2 steps)
