+++ package.json
@@ -34,7 +34,8 @@
         "copy:prepare-binskim": "node ./pipeline/scripts/prepare-binskim-symbols.js",
         "copyright:check": "license-check-and-add check -f copyright-header.config.json",
         "copyright:fix": "license-check-and-add add -f copyright-header.config.json",
-        "download:electron-mirror": "node ./pipeline/scripts/download-electron-mirror.js",
+        "download:electron-mirror:binaries": "node ./pipeline/scripts/download-electron-mirror-binaries.js",
+        "download:electron-mirror:symbols": "node ./pipeline/scripts/download-electron-mirror-symbols.js",
         "fastpass": "npm-run-all --print-label scss:build --parallel copyright:check lint:check format:check null:check && grunt ada-cat",
         "fastpass:fix": "npm-run-all --print-label --serial scss:clean copyright:fix lint:fix format:fix",
         "format:check": "prettier --config prettier.config.js --check \"**/*\"",
